sanlock: lockspace callbacks
sanlock_reg_lockspace() returns an fd that the caller should
use in poll(2), and when ready should call sanlock_get_callback()
to read the callback.

Currently, the only callback type is host message.  When a host
receives a message from another host, the host message is sent
to the callback fd.

Signed-off-by: David Teigland <teigland@redhat.com>