reset: Require lockspace name for local deamon operations
sanlk-reset required lockspace name for the reset command, but did not
require it for reg, end, and clear commands. Running a comment without
did nothing and exit with zero exit code.

Now a lockspace name is required for all commands and the tool exit with
exit code 2 if not available.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>