Reduce logging noise to stderr and syslog
When reseting another host, we used to create lot of noise by logging
the result of any lockspace check every 2 seconds. This info is useless
for the user of this command, but may be valuable for debugging. So now
we log everything when using --debug-mode, and only state changes
otherwise.

Some log messages were too verbose and technical. Now log_debug messages
are verbose and technical, but log_info messages are trying to be user
friendly.

Previously we logged few messages per lockspace to syslog for each reset
operation. This is unwanted noise on a machine performing a reset,
specially when using 10's of lockspaces. Now only warning and errors
are logged to syslog. We assume that a program running this tool will
log its own logs if needed for each reset operation.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>